requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
openai
python-dotenv